ref(utils): Move duration related into the same folder
