+++ static/app/components/events/interfaces/spans/aggregateSpanDetail.tsx
@@ -4,7 +4,8 @@ import {t} from 'sentry/locale';
 import {space} from 'sentry/styles/space';
 import {Organization} from 'sentry/types';
 import {AggregateEventTransaction} from 'sentry/types/event';
-import {formatPercentage, getDuration} from 'sentry/utils/formatters';
+import getDuration from 'sentry/utils/duration/getDuration';
+import {formatPercentage} from 'sentry/utils/formatters';
 import {QuickTraceEvent, TraceError} from 'sentry/utils/performance/quickTrace/types';
 
 import {AggregateSpanType, ParsedTraceType} from './types';
