+++ static/app/components/quickTrace/styles.tsx
@@ -8,7 +8,7 @@ import MenuItem from 'sentry/components/menuItem';
 import Tag, {Background} from 'sentry/components/tag';
 import Truncate from 'sentry/components/truncate';
 import {space} from 'sentry/styles/space';
-import {getDuration} from 'sentry/utils/formatters';
+import getDuration from 'sentry/utils/duration/getDuration';
 import {QuickTraceEvent} from 'sentry/utils/performance/quickTrace/types';
 
 export const SectionSubtext = styled('div')`
