+++ static/app/utils/formatters.spec.tsx
@@ -6,8 +6,6 @@ import {
   formatPercentage,
   formatRate,
   formatSecondsToClock,
-  getDuration,
-  getExactDuration,
   MONTH,
   parseClockToSeconds,
   userDisplayName,
@@ -19,92 +17,18 @@ describe('getDuration()', function () {
     expect(formatSecondsToClock(0)).toBe('00:00');
     expect(formatSecondsToClock(0.001)).toBe('00:00.001');
     expect(formatSecondsToClock(0.01)).toBe('00:00.010');
-    expect(getDuration(0.1)).toBe('100 milliseconds');
-    expect(getDuration(0.1, 2)).toBe('100.00 milliseconds');
-    expect(getDuration(1)).toBe('1 second');
-    expect(getDuration(2)).toBe('2 seconds');
-    expect(getDuration(65)).toBe('1 minute');
-    expect(getDuration(122)).toBe('2 minutes');
-    expect(getDuration(3720)).toBe('1 hour');
-    expect(getDuration(36000)).toBe('10 hours');
-    expect(getDuration(86400)).toBe('1 day');
-    expect(getDuration(86400 * 2)).toBe('2 days');
-    expect(getDuration(604800)).toBe('1 week');
-    expect(getDuration(604800 * 4)).toBe('4 weeks');
-    expect(getDuration(2629800)).toBe('1 month');
-    expect(getDuration(604800 * 12)).toBe('3 months');
   });
 
   it('should format negative durations', function () {
     expect(formatSecondsToClock(0)).toBe('00:00');
     expect(formatSecondsToClock(-0.001)).toBe('00:00.001');
     expect(formatSecondsToClock(-0.01)).toBe('00:00.010');
-    expect(getDuration(-0.1)).toBe('-100 milliseconds');
-    expect(getDuration(-0.1, 2)).toBe('-100.00 milliseconds');
-    expect(getDuration(-1)).toBe('-1 second');
-    expect(getDuration(-2)).toBe('-2 seconds');
-    expect(getDuration(-65)).toBe('-1 minute');
-    expect(getDuration(-122)).toBe('-2 minutes');
-    expect(getDuration(-3720)).toBe('-1 hour');
-    expect(getDuration(-36000)).toBe('-10 hours');
-    expect(getDuration(-86400)).toBe('-1 day');
-    expect(getDuration(-86400 * 2)).toBe('-2 days');
-    expect(getDuration(-604800)).toBe('-1 week');
-    expect(getDuration(-604800 * 4)).toBe('-4 weeks');
-    expect(getDuration(-2629800)).toBe('-1 month');
-    expect(getDuration(-604800 * 12)).toBe('-3 months');
-  });
-
-  it('should format numbers and abbreviate units', function () {
-    expect(getDuration(0, 2, true)).toBe('0.00ms');
-    expect(getDuration(0, 0, true)).toBe('0ms');
-    expect(getDuration(0.1, 0, true)).toBe('100ms');
-    expect(getDuration(0.1, 2, true)).toBe('100.00ms');
-    expect(getDuration(1, 2, true)).toBe('1.00s');
-    expect(getDuration(122, 0, true)).toBe('2min');
-    expect(getDuration(3600, 0, true)).toBe('1hr');
-    expect(getDuration(86400, 0, true)).toBe('1d');
-    expect(getDuration(86400 * 2, 0, true)).toBe('2d');
-    expect(getDuration(604800, 0, true)).toBe('1wk');
-    expect(getDuration(604800 * 2, 0, true)).toBe('2wk');
-    expect(getDuration(2629800, 0, true)).toBe('1mo');
-    expect(getDuration(604800 * 12, 0, true)).toBe('3mo');
-  });
-
-  it('should format numbers and abbreviate units with one letter', function () {
-    expect(getDuration(0, 2, false, true)).toBe('0.00ms');
-    expect(getDuration(0, 0, false, true)).toBe('0ms');
-    expect(getDuration(0.1, 0, false, true)).toBe('100ms');
-    expect(getDuration(0.1, 2, false, true)).toBe('100.00ms');
-    expect(getDuration(1, 2, false, true)).toBe('1.00s');
-    expect(getDuration(122, 0, false, true)).toBe('2m');
-    expect(getDuration(3600, 0, false, true)).toBe('1h');
-    expect(getDuration(86400, 0, false, true)).toBe('1d');
-    expect(getDuration(86400 * 2, 0, false, true)).toBe('2d');
-    expect(getDuration(604800, 0, false, true)).toBe('1w');
-    expect(getDuration(604800 * 2, 0, false, true)).toBe('2w');
-    expect(getDuration(2629800, 0, false, true)).toBe('4w');
-    expect(getDuration(604800 * 12, 0, false, true)).toBe('12w');
   });
 
   it('should format negative durations with absolute', function () {
     expect(formatSecondsToClock(0)).toBe('00:00');
     expect(formatSecondsToClock(-0.001)).toBe('00:00.001');
     expect(formatSecondsToClock(-0.01)).toBe('00:00.010');
-    expect(getDuration(-0.1, 0, false, false, true)).toBe('100 milliseconds');
-    expect(getDuration(-0.1, 2, false, false, true)).toBe('100.00 milliseconds');
-    expect(getDuration(-1, 0, false, false, true)).toBe('1 second');
-    expect(getDuration(-2, 0, false, false, true)).toBe('2 seconds');
-    expect(getDuration(-65, 0, false, false, true)).toBe('1 minute');
-    expect(getDuration(-122, 0, false, false, true)).toBe('2 minutes');
-    expect(getDuration(-3720, 0, false, false, true)).toBe('1 hour');
-    expect(getDuration(-36000, 0, false, false, true)).toBe('10 hours');
-    expect(getDuration(-86400, 0, false, false, true)).toBe('1 day');
-    expect(getDuration(-86400 * 2, 0, false, false, true)).toBe('2 days');
-    expect(getDuration(-604800, 0, false, false, true)).toBe('1 week');
-    expect(getDuration(-604800 * 4, 0, false, false, true)).toBe('4 weeks');
-    expect(getDuration(-2629800, 0, false, false, true)).toBe('1 month');
-    expect(getDuration(-604800 * 12, 0, false, false, true)).toBe('3 months');
   });
 });
 
@@ -321,37 +245,5 @@ describe('userDisplayName', function () {
     // @ts-expect-error
     expect(userDisplayName()).toEqual('Unknown author');
     expect(userDisplayName({})).toEqual('Unknown author');
-  });
-});
-
-describe('getExactDuration', () => {
-  it('should provide default value', () => {
-    expect(getExactDuration(0)).toEqual('0 milliseconds');
-  });
-
-  it('should format in the right way', () => {
-    expect(getExactDuration(2.030043848568126)).toEqual('2 seconds 30 milliseconds');
-    expect(getExactDuration(0.2)).toEqual('200 milliseconds');
-    expect(getExactDuration(13)).toEqual('13 seconds');
-    expect(getExactDuration(60)).toEqual('1 minute');
-    expect(getExactDuration(121)).toEqual('2 minutes 1 second');
-    expect(getExactDuration(234235435)).toEqual(
-      '387 weeks 2 days 1 hour 23 minutes 55 seconds'
-    );
-  });
-
-  it('should format negative durations', () => {
-    expect(getExactDuration(-2.030043848568126)).toEqual('-2 seconds -30 milliseconds');
-    expect(getExactDuration(-0.2)).toEqual('-200 milliseconds');
-    expect(getExactDuration(-13)).toEqual('-13 seconds');
-    expect(getExactDuration(-60)).toEqual('-1 minute');
-    expect(getExactDuration(-121)).toEqual('-2 minutes -1 second');
-    expect(getExactDuration(-234235435)).toEqual(
-      '-387 weeks -2 days -1 hour -23 minutes -55 seconds'
-    );
-  });
-
-  it('should abbreviate label', () => {
-    expect(getExactDuration(234235435, true)).toEqual('387wk 2d 1hr 23min 55s');
   });
 });
