+++ static/app/utils/performance/histogram/utils.tsx
@@ -1,5 +1,5 @@
 import {ColumnType} from 'sentry/utils/discover/fields';
-import {getDuration} from 'sentry/utils/formatters';
+import getDuration from 'sentry/utils/duration/getDuration';
 
 import {HistogramData} from './types';
