+++ static/app/views/settings/project/projectKeys/details/keyRateLimitsForm.tsx
@@ -16,7 +16,7 @@ import {t, tct, tn} from 'sentry/locale';
 import {space} from 'sentry/styles/space';
 import {Organization} from 'sentry/types';
 import {defined} from 'sentry/utils';
-import {getExactDuration} from 'sentry/utils/formatters';
+import getExactDuration from 'sentry/utils/duration/getExactDuration';
 import {ProjectKey} from 'sentry/views/settings/project/projectKeys/types';
 
 const PREDEFINED_RATE_LIMIT_VALUES = [
